requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.3
flask-cors
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
nulltype==2.3.1
plaid-python==38.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests
six==1.17.0
openai
supabase
urllib3==2.6.3
Werkzeug==3.1.6
twilio>=9.0.0
requests>=2.31.0